==> 5 make features/3 patent.ipynb <==
# patent data에서 필요한 firm-lv 변수 뽑기

### 1. patent에 gvkey, country, year 붙이기
- gvkey, country: assignee
- year: patent_class

### 2. 변수 제작
- fwd, bwd cit (sum, avg)
- self citation intensity = self / 전체 (fwd, bwd)
- size
- country: firm fixed effect 하면 사라질건데... 일단 구해놓자 중복 많네 귀찮게... 나중에 수작업으로 선별해야
- class로 구한 (normalized) herfindahl index

### 1. patent info - gvkey, country, year
- year 붙이기 (patent_class, patent_info join)  
__=> patent_id, year, info__  
- patent_id -> assignee -> gvkey (patent_assignee join, assignee_gvkey join)   
__=> gvkey, year, info, assignee__
- gvkey, assignee만 분리 후, assignee -> raw_assignee -> country (assignee_raw join, assignee join)
__=> gvkey, country__

import pandas as pd
import numpy as np
import csv
import os
import math

os.chdir('E:/apps/db_table')

__1. patent_id, year, info__

bio = pd.read_csv('patent_class_BIO.csv')
ee = pd.read_csv('patent_class_EE&IT.csv')

bio = bio.loc[:,['patent_id','year']]
ee = ee.loc[:,['patent_id','year']]

df = pd.read_csv('patent_info.csv')

bio = bio.merge(df,how='left',on='patent_id')
ee = ee.merge(df,how='left',on='patent_id')

len(bio), len(ee)

list(bio)

__2. patent_id -> gvkey__

df = pd.read_csv('patent_assignee.csv')

bio = bio.merge(df, how='left', on='patent_id')
ee = ee.merge(df, how='left', on='patent_id')

len(bio), len(ee)

df = pd.read_csv('assignee_gvkey_BIO.csv')
df1 = pd.read_csv('assignee_gvkey_EE&IT.csv')

bio = bio.merge(df, how='left', on='assignee')
ee = ee.merge(df1, how='left', on='assignee')

len(bio), len(ee)

bio.dropna(how='any',inplace=True)  #drop patents which don't have gvkey of assignee
ee.dropna(how='any',inplace=True)

len(bio), len(ee)

bio.to_csv('bio_temp.csv',index=False)
ee.to_csv('ee_temp.csv',index=False)

__3. gvkey -> country__

bio = pd.read_csv('bio_temp.csv')
ee = pd.read_csv('ee_temp.csv')

biocon = bio.loc[:,['gvkey','assignee']]
eecon = ee.loc[:,['gvkey','assignee']]

biocon.drop_duplicates(inplace=True)
eecon.drop_duplicates(inplace=True)

len(biocon), len(eecon)

df = pd.read_csv('assignee.csv')
df = df.loc[:,['country_code','pdpass']]
df.dropna(how='any',inplace=True)
df.drop_duplicates(inplace=True)

len(df)

pdp = pd.read_csv('pdpass.csv')

df = df.merge(pdp,how='left',on='pdpass')

biocon = biocon.merge(df, how='left', on='assignee')
eecon = eecon.merge(df, how='left', on='assignee')

biocon = biocon.loc[:,['gvkey','country_code']]
eecon = eecon.loc[:,['gvkey','country_code']]

biocon.drop_duplicates(inplace=True)
eecon.drop_duplicates(inplace=True)

len(biocon), len(eecon)

biocon.to_csv('country_bio.csv',index=False)
eecon.to_csv('country_EE&IT.csv',index=False)

### 2 .변수 제작
- groupby gvkey, year
- size (count)
- claim, fwd, bwd, self, self_fwd cit (avg) => 그 후 self citation intensity = self / 전체 (fwd, bwd)
- Herf index

list(bio)

del bio['assignee']
del ee['assignee']

__1. size__

biosize = bio.loc[:,['gvkey','year']]
eesize = ee.loc[:,['gvkey','year']]

biosize['pat_size'] = ''
eesize['pat_size'] = ''

biosize = biosize.groupby(['gvkey','year'], as_index=False).agg('count')
eesize = eesize.groupby(['gvkey','year'], as_index=False).agg('count')

len(biosize), len(eesize)

biosize.iloc[0:10,:]

__2. citation__

bio = bio.rename(columns = {'num_us_citations':'bwd','num_times_cited':'fwd','num_self_citations':'self',
                            'num_self_future_citations':'self_fwd'})
ee = ee.rename(columns = {'num_us_citations':'bwd','num_times_cited':'fwd','num_self_citations':'self',
                            'num_self_future_citations':'self_fwd'})

biocit = bio.loc[:,['gvkey','year','num_claims','bwd','fwd','self','self_fwd']]
eecit = ee.loc[:,['gvkey','year','num_claims','bwd','fwd','self','self_fwd']]

biocit = biocit.groupby(['gvkey','year'], as_index=False).mean()
eecit = eecit.groupby(['gvkey','year'], as_index=False).mean()

biocit['self_ratio'] = biocit['self'] / biocit['bwd']
biocit['self_ratio_fwd'] = biocit['self_fwd'] / biocit['fwd']

eecit['self_ratio'] = eecit['self'] / eecit['bwd']
eecit['self_ratio_fwd'] = eecit['self_fwd'] / eecit['fwd']

len(biocit), len(eecit)

biocit.iloc[0:10,:]

bio = biosize.merge(biocit,how='left',on=['gvkey','year'])
ee = eesize.merge(eecit,how='left',on=['gvkey','year'])

len(bio), len(ee)

bio1 = bio.drop(['self_ratio','self_ratio_fwd'], axis=1)
ee1 = ee.drop(['self_ratio','self_ratio_fwd'], axis=1)

bio1.to_csv('E:/apps/feature/patent_bio.csv',index=False)
ee1.to_csv('E:/apps/feature/patent_EE&IT.csv',index=False)

bio.isnull().any()

bio.dropna(how='any',inplace=True)
ee.dropna(how='any',inplace=True)

len(bio), len(ee)

bio.to_csv('E:/apps/feature/patent_self_bio.csv',index=False)
ee.to_csv('E:/apps/feature/patent_self_EE&IT.csv',index=False)

__3. herfindhal index__

import pandas as pd
import numpy as np
import csv
import os
import math

os.chdir('E:/apps/appcit measure')

header = ['gvkey','year','herf_cls']

def get_herf(ind,y_start,y_end):
    fname = 'E:/apps/feature/cls_herf_' + ind + '.csv'
    with open(fname,'wb') as f:
        writer = csv.writer(f)
        writer.writerow(header)
        
        for y in range(y_start,1+y_end):   
            cls = 'class_vector_' + ind + '_' + str(y) +'.csv'
            df = pd.read_csv(cls)            
            
            for i in range(0,len(df)):
                row = []
                row.append(df.iloc[i,0])
                row.append(y)
                herf = round(sum([x**2 for x in df.iloc[i,1:]]),3)
                row.append(herf)
                
                writer.writerow(row)

get_herf('bio',2005,2012)
get_herf('EE&IT',2005,2012)
