==> 5 make features/5 final csv.ipynb <==
# feature들 합쳐서 분석 가능하게 만들기

### 1. X (IV, control) 다 때려넣기
- 기준 고르기: 원래라면 compu, 걍 app_cit으로 할까? app_cit 한 번이라도 하거나 받은 애들 조사할거니까
- na처리: drop / 0 / 평균 잘 생각

### 2. DV 고르기 (3, 4, 5년 짜리, 공수) 

### 3. dummy 제작 (firm fixed, year fixed)

### 4. 보수공사

#-*- coding: utf-8 -*-
import os
import pandas as pd
import numpy as np

os.chdir('E:/apps/feature')

### 1. X 만들기
__app cit 기준__

bio = pd.read_csv('appcit_bio.csv')
ee = pd.read_csv('appcit_EE&IT.csv')

len(bio), len(ee)

__patent 붙이기__

df = pd.read_csv('patent_bio.csv')
df1 = pd.read_csv('patent_EE&IT.csv')

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

len(bio), len(ee)

bio.isnull().any()

ee.isnull().any()

df = pd.read_csv('patent_self_bio.csv')
df1 = pd.read_csv('patent_self_EE&IT.csv')

df = df.loc[:,['gvkey','year','self_ratio','self_ratio_fwd']]
df1 = df1.loc[:,['gvkey','year','self_ratio','self_ratio_fwd']]

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

ee.isnull().any()

bio.isnull().any()

여기는 얘네 둘 쓰려면 __dropna__ 필요

__compustat 붙이기__

df = pd.read_csv('compu_bio.csv')
df1 = pd.read_csv('compu_EE&IT.csv')

list(df)

df = df.loc[:,['gvkey','fyear','roa','margin','rdi','mkt_4','mkt_3']]
df1 = df1.loc[:,['gvkey','fyear','roa','margin','rdi','mkt_4','mkt_3']]

df = df.rename(columns ={'fyear':'year'})
df1 = df1.rename(columns ={'fyear':'year'})

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

ee.isnull().any()

bio.isnull().any()

여기는 얘네 둘 쓰려면 __average__ 때려버리자

__tech concentration 붙이기__

df = pd.read_csv('cls_herf_bio.csv')
df1 = pd.read_csv('cls_herf_EE&IT.csv')

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

ee.isnull().any()

bio.isnull().any()

bio.to_csv('X_bio_na.csv',index=False)
ee.to_csv('X_EE&IT_na.csv',index=False)

### 2. DV 고르기
- 공수 붙이기: 공수 outer join, X에 left join 후 fillna
- 3, 4, 5년은 따로

bio = pd.read_csv('X_bio_na.csv')
ee = pd.read_csv('X_EE&IT_na.csv')

ba = pd.read_csv('lit_a_3_bio.csv')
bd = pd.read_csv('lit_d_3_bio.csv')

ea = pd.read_csv('lit_a_3_EE&IT.csv')
ed = pd.read_csv('lit_d_3_EE&IT.csv')

print len(ba), len(bd), len(ea), len(ed)

b = ba.merge(bd,how='outer',on=['gvkey','year'])
e = ea.merge(ed,how='outer',on=['gvkey','year'])

len(b), len(e)

len(bio), len(ee)

bio = bio.merge(b,how='left',on=['gvkey','year'])
ee = ee.merge(e,how='left',on=['gvkey','year'])

bio.loc[:, ['lit_a_3', 'lit_d_3']] = bio.loc[:, ['lit_a_3', 'lit_d_3']].fillna(0)
ee.loc[:, ['lit_a_3', 'lit_d_3']] = ee.loc[:, ['lit_a_3', 'lit_d_3']].fillna(0)

ee.isnull().any()

bio.isnull().any()

### 3. dummy들 만들기

bio = pd.get_dummies(bio, dummy_na=True, columns=['gvkey','year'])
ee = pd.get_dummies(ee, dummy_na=True, columns=['gvkey','year'])

bio.to_csv('X_bio_na_dum_3.csv',index=False)
ee.to_csv('X_EE&IT_na_dum_3.csv',index=False)

### 4. 보수공사
- compustat에서 빠진 변수
- patent에서 빠진 변수

__compustat에서 빠진 변수__

bio = pd.read_csv('X_bio_na.csv')
ee = pd.read_csv('X_EE&IT_na.csv')

df = pd.read_csv('compu_bio.csv')
df1 = pd.read_csv('compu_EE&IT.csv')

df = df.loc[:,['gvkey','fyear','xrd','emp']]
df1 = df1.loc[:,['gvkey','fyear','xrd','emp']]

df = df.rename(columns ={'fyear':'year'})
df1 = df1.rename(columns ={'fyear':'year'})

def emp(x):
    if x>0:
        return np.log(x*1000)
    else:
        return np.nan
    
def rnd(x):
    if x>0:
        return np.log(x*1000000)
    else:
        return np.nan

df['emp'] = df['emp'].apply(emp)
df['rnd'] = df['xrd'].apply(rnd)
del df['xrd']

df1['emp'] = df1['emp'].apply(emp)
df1['rnd'] = df1['xrd'].apply(rnd)
del df1['xrd']

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

bio.isnull().any()

ee.isnull().any()

bio.to_csv('X_bio_na1.csv',index=False)
ee.to_csv('X_EE&IT_na1.csv',index=False)

len(bio),len(ee)

bio.dropna(inplace=True)
ee.dropna(inplace=True)

len(bio),len(ee)

bio.to_csv('X_bio_dropna.csv',index=False)
ee.to_csv('X_EE&IT_dropna.csv',index=False)

__patent에서 빠진 변수__

bio = pd.read_csv('X_bio_na1.csv')
ee = pd.read_csv('X_EE&IT_na1.csv')

len(bio), len(ee)

df = pd.read_csv('patcit_bio.csv')
df1 = pd.read_csv('patcit_EE&IT.csv')

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

df = pd.read_csv('patent_stock_bio.csv')
df1 = pd.read_csv('patent_stock_EE&IT.csv') 

bio = bio.merge(df,how='left',on=['gvkey','year'])
ee = ee.merge(df1,how='left',on=['gvkey','year'])

len(bio),len(ee)

bio.to_csv('X_bio_na2.csv',index=False)
ee.to_csv('X_EE&IT_na2.csv',index=False)

bio.dropna(inplace=True)
ee.dropna(inplace=True)

len(bio),len(ee)

bio.to_csv('X_bio_dropna.csv',index=False)
ee.to_csv('X_EE&IT_dropna.csv',index=False)

__appcit / patcit 만들기__
- non-weighted: edge_list groupby count 
- 기업별 app/pat 비율 구하기
- X 에다가 붙이기

#-*- coding: utf-8 -*-
import os
import pandas as pd
import numpy as np

os.chdir('E:/apps')

a_bio = pd.read_csv('app_cit_edge_bio.csv')
a_ee = pd.read_csv('app_cit_edge_EE&IT.csv')

p_bio = pd.read_csv('pat_cit_edge_BIO.csv')
p_ee = pd.read_csv('pat_cit_edge_EE&IT.csv')

a_bio = a_bio.groupby(['gvkey','year']).size().reset_index(name='a_counts')
a_ee = a_ee.groupby(['gvkey','year']).size().reset_index(name='a_counts')

p_bio = p_bio.groupby(['gvkey','year']).size().reset_index(name='p_counts')
p_ee = p_ee.groupby(['gvkey','year']).size().reset_index(name='p_counts')

len(a_bio), len(p_bio), len(a_ee), len(p_ee)

bio = p_bio.merge(a_bio,how='left',on=['gvkey','year']).fillna(0)
ee = p_ee.merge(a_ee,how='left',on=['gvkey','year']).fillna(0)

len(bio), len(ee)

bio['ap_ratio'] = bio['a_counts'] / bio['p_counts']
ee['ap_ratio'] = ee['a_counts'] / ee['p_counts']

print sum(ee['ap_ratio'])/len(ee), sum(bio['ap_ratio'])/len(bio)

os.chdir('E:/apps/feature')

bio_x = pd.read_csv('X_bio_avgna.csv') 
ee_x = pd.read_csv('X_EE&IT_avgna.csv') 

len(bio_x), len(ee_x)

bio = bio.loc[:,['gvkey','year','ap_ratio']]
ee = ee.loc[:,['gvkey','year','ap_ratio']]

bio_x = bio_x.merge(bio,how='left',on=['gvkey','year'])
ee_x=ee_x.merge(ee,how='left',on=['gvkey','year'])

len(bio_x), len(ee_x)

bio_x.to_csv('X_bio_avgna_1.csv',index=False)
ee_x.to_csv('X_EE&IT_avgna_1.csv',index=False) 

bio_x.dropna(inplace=True)
ee_x.dropna(inplace=True)

len(bio_x),len(ee_x)

bio_x.to_csv('X_bio_dropna_1.csv',index=False)
ee_x.to_csv('X_EE&IT_dropna_1.csv',index=False)
