==> 5 make features/7. firm-to-firm csv.ipynb <==
# firm-firm-year csv 만들기

### 1. firm i, j 변수 그대로 사용
- X_i, X_j join (left join)
- forward 빼고 붙이기

### 2. firm i 변수, firm j ratio 사용
- X_i, X_j로 X_ij 만들기 (X_i / X_j)
- X_j 제거

import pandas as pd
import numpy as np
import csv
import os
import math

os.chdir('E:/apps/feature_ffy')

### 1. firm i,j 변수 그대로
** 쓸모없는 변수는 엑셀로 미리 제거 **

bio = pd.read_csv('X_bio_mvars.csv')
ee = pd.read_csv('X_EE&IT_mvars.csv')

bx = pd.read_csv('X_bio.csv')
ex = pd.read_csv('X_EE&IT.csv')

len(bio), len(ee)

bio = bio.merge(bx,how='left',on=['gvkey','year'])
ee = ee.merge(ex,how='left',on=['gvkey','year'])

len(bio), len(ee)

bx.rename(columns={"gvkey": "gvkey_cited"},inplace=True)
ex.rename(columns={"gvkey": "gvkey_cited"},inplace=True)

list(bx)

bio = bio.merge(bx,how='left',on=['gvkey_cited','year'],suffixes=('_i', '_j'))
ee = ee.merge(ex,how='left',on=['gvkey_cited','year'],suffixes=('_i', '_j'))

len(bio), len(ee)

list(bio)

bio.dropna(how='any',inplace=True)
ee.dropna(how='any',inplace=True)
len(bio), len(ee)

bio.to_csv('X_bio_ij.csv',index=False)
ee.to_csv('X_EE&IT_ij.csv',index=False)

### 2. firm i, firm j ratio

len(list(bio)), list(bio).index('num_claims_i'), list(bio).index('num_claims_j')

varname = list(bio)[8]+'j'
bio[varname] = bio[list(bio)[8]]/bio[list(bio)[19]]
list(bio)[30]

del bio[list(bio)[19]]

len(list(bio)),list(bio)[19],list(bio)[29]

bio = pd.read_csv('X_bio_ij.csv')

def get_ratio(df):
    for i in range(8,19):
        varname = list(df)[i]+'j'
        df[varname] = df[list(df)[i]]/df[list(df)[19]]
        del df[list(df)[19]]
    
    return df

get_ratio(bio)
get_ratio(ee)

list(bio)

bio.to_csv('X_bio_ij_ratio.csv',index=False)
ee.to_csv('X_EE&IT_ij_ratio.csv',index=False)
